+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 
 * **cleanup frequency detector : remove all coverage related stuff**
 [1239aca520116b0](https://github.com/cylab/mark/commit/1239aca520116b0) Thibault Debatty *2020-10-23 12:03:00*
+* **Normalization only if weights sum is != to 1.0. Avoid a rounded result that fails the tests**
+[37ddf79ab493033](https://github.com/cylab/mark/commit/37ddf79ab493033) Alex *2020-10-23 09:35:45*
+* **Add normalization vector in OWA to avoid error if weights sum != 1**
+[b971d97d281dab6](https://github.com/cylab/mark/commit/b971d97d281dab6) Alex *2020-10-23 09:25:41*
+* **Add normlization vector to avoid error if weights sum is different to 1**
+[8ef632cecd6f3b0](https://github.com/cylab/mark/commit/8ef632cecd6f3b0) Alex *2020-10-23 09:23:08*
 * **updated changelog**
 [7880b878995785b](https://github.com/cylab/mark/commit/7880b878995785b) Thibault Debatty *2020-10-23 09:13:24*
 * **add link to online javadoc**